Throw configuration error when `tar` is not available
