+++ lib/setup-codeql.js
@@ -463,6 +463,9 @@ function getCanonicalToolcacheVersion(cliVersion, bundleVersion, logger) {
  * @returns the path to the extracted bundle, and the version of the tools
  */
 async function setupCodeQLBundle(toolsInput, apiDetails, tempDir, variant, features, defaultCliVersion, logger) {
+    if (!(await util.isBinaryAccessible("tar", logger))) {
+        throw new util.ConfigurationError("Could not find tar in PATH, so unable to extract CodeQL bundle.");
+    }
     const zstdAvailability = await tar.isZstdAvailable(logger);
     const source = await getCodeQLSource(toolsInput, defaultCliVersion, apiDetails, variant, zstdAvailability.available, logger);
     let codeqlFolder;
