+++ lib/util.js
@@ -69,12 +69,14 @@ exports.checkActionVersion = checkActionVersion;
 exports.cloneObject = cloneObject;
 exports.checkSipEnablement = checkSipEnablement;
 exports.cleanUpGlob = cleanUpGlob;
+exports.isBinaryAccessible = isBinaryAccessible;
 const fs = __importStar(require("fs"));
 const os = __importStar(require("os"));
 const path = __importStar(require("path"));
 const util_1 = require("util");
 const core = __importStar(require("@actions/core"));
 const exec = __importStar(require("@actions/exec/lib/exec"));
+const safe_which_1 = require("@chrisgavin/safe-which");
 const check_disk_space_1 = __importDefault(require("check-disk-space"));
 const del_1 = __importDefault(require("del"));
 const get_folder_size_1 = __importDefault(require("get-folder-size"));
@@ -924,6 +926,17 @@ async function cleanUpGlob(glob, name, logger) {
     }
     catch (e) {
         logger.warning(`Failed to clean up ${name}: ${e}.`);
+    }
+}
+async function isBinaryAccessible(binary, logger) {
+    try {
+        await (0, safe_which_1.safeWhich)(binary);
+        logger.debug(`Found ${binary}.`);
+        return true;
+    }
+    catch (e) {
+        logger.debug(`Could not find ${binary}: ${e}`);
+        return false;
     }
 }
 //# sourceMappingURL=util.js.map
